add camera infor provider
